% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprocess_function.R
\name{sprocess$fitmodel}
\alias{sprocess$fitmodel}
\title{A spatial simulation function}
\usage{
sprocess$fitmodel
}
\arguments{
\item{x1}{a vector of longitudes}

\item{x2}{a vector of latitudes}

\item{sigmasq.true}{true sigma squared value}

\item{tausq.true}{true tau squared value}

\item{phi.true}{true phi value}

\item{phi.upper}{upper bound of a discrete uniform prior on phi}

\item{tausq.upper}{upper bound of a discrete uniform prior on tau squared}
}
\description{
This function allows you to simulate a spatial process by specifying the known parameters.
}
\examples{
fitmodel(x1,x2,10,0.15,0.5,0.7,1)
sprocess_function()
}
\keyword{spatial}
